tratamento de exceçÂões

+++ src/main/java/com/crudcliente/services/ClienteService.java
@@ -3,10 +3,15 @@ package com.crudcliente.services;
 import com.crudcliente.dto.ClienteDTO;
 import com.crudcliente.entities.Cliente;
 import com.crudcliente.repositories.ClienteRepository;
+import com.crudcliente.services.exceptions.DatabaseException;
+import com.crudcliente.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +22,8 @@ public class ClienteService {
 
     @Transactional(readOnly = true)
     public ClienteDTO findById(Long id){
-        Cliente cliente = repository.findById(id).get();
+        Cliente cliente = repository.findById(id).orElseThrow(
+                ()-> new ResourceNotFoundException("Recurso não encontrado"));
         return new ClienteDTO(cliente);
     }
 
@@ -37,16 +43,29 @@ public class ClienteService {
 
     @Transactional
     public ClienteDTO update(Long id, ClienteDTO dto){
-        Cliente entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClienteDTO(entity);
+        try {
+            Cliente entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClienteDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
